# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  baby-backend_internal_models.Baby:
    properties:
      avatar:
        type: string
      birth_height:
        type: string
      birth_weight:
        type: string
      birthday:
        description: '"2006-01-02"'
        type: string
      blood_type:
        type: string
      created_at:
        type: string
      gender:
        description: '"boy" | "girl"'
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  baby-backend_internal_services.AuthResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/baby-backend_internal_services.UserSafeView'
    type: object
...
  baby-backend_internal_services.CreateBabyInput:
    properties:
      avatar:
        type: string
      birth_height:
        type: string
      birth_weight:
        type: string
      birthday:
        type: string
      blood_type:
        type: string
      gender:
        enum:
        - boy
        - girl
        type: string
      name:
        type: string
      notes:
        type: string
    required:
    - birthday
    - gender
    - name
    type: object
...
  baby-backend_internal_services.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  baby-backend_internal_services.RegisterInput:
    properties:
      email:
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  baby-backend_internal_services.UpdateBabyInput:
    properties:
      avatar:
        type: string
      birth_height:
        type: string
      birth_weight:
        type: string
      birthday:
        type: string
      blood_type:
        type: string
      gender:
        type: string
      name:
        type: string
      notes:
        type: string
    type: object
  baby-backend_internal_services.UserSafeView:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  baby-backend_internal_utils.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 宝宝时光 · 家庭馆 — 宝宝成长记录与家庭共享后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com
  title: Baby Backend API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 支持用户名或邮箱登录，成功返回 JWT token。
      parameters:
      - description: 登录信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_services.AuthResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      summary: 用户登录
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 创建新帐号，返回 JWT token。用户名和邮箱均需唯一。
      parameters:
      - description: 注册信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_services.AuthResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      summary: 用户注册
      tags:
      - Auth
  /babies:
    get:
      description: 获取当前用户下的所有宝宝。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.Baby'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 宝宝列表
      tags:
      - Baby
    post:
      consumes:
      - application/json
      description: 创建一个新的宝宝档案。
      parameters:
      - description: 宝宝信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateBabyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.Baby'
              type: object
      security:
      - BearerAuth: []
      summary: 添加宝宝
      tags:
      - Baby
  /babies/{id}:
    delete:
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除宝宝
      tags:
      - Baby
    get:
      description: 按 ID 获取单个宝宝。
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.Baby'
              type: object
      security:
      - BearerAuth: []
      summary: 宝宝详情
      tags:
      - Baby
    put:
      consumes:
      - application/json
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.UpdateBabyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.Baby'
              type: object
      security:
      - BearerAuth: []
      summary: 编辑宝宝
      tags:
      - Baby
...
  /profile:
    get:
      description: 返回已登录用户的 id 和 username。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 当前用户信息
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"